yarn install only in the webpacker container

+++ scripts/docker/entrypoint.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 set -e
 set -x
-scripts/docker/setup.sh
 
-# Remove a potentially pre-existing server.pid for Rails.
-rm -f /app/tmp/pids/server.pid
+if [[ $@ =~ 'webpack' ]]
+then
+  yarn install --frozen-lockfile
+else
+  bundle check || bundle install || true
+  # Remove a potentially pre-existing server.pid for Rails.
+  rm -f /app/tmp/pids/server.pid
+fi
 
-# yarn install
 # Then exec the container's main process (what's set as CMD in the Dockerfile).
 exec "$@"
